requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
numpy>=1.21.0
PyAudio>=0.2.11
webrtcvad>=2.0.10
PyQt6>=6.5.0
sounddevice>=0.4.6
soundfile>=0.10.3   